% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{console.plot-pkg}
\alias{console.plot-pkg}
\alias{console.plot-pkg-package}
\title{console.plot: View R plots in the console}
\description{
console.plot: View R plots in the console
}
\details{
Create text-based plots.
}
